+++ docs/introduction/version_history.rst
@@ -11,7 +11,7 @@ Version 0.2.0 Alpha (Not yet released)
 * Cleanup m4 files (`Issue #28 <https://github.com/libattachsql/libattachsql/issues/28>`_)
 * Fixed in_query flag status (`Issue #31 <https://github.com/libattachsql/libattachsql/issues/31>`_)
 * Add row buffering support (`Issue #8 <https://github.com/libattachsql/libattachsql/issues/8>`_)
-
+* Connect on first query is now non-blocking (`Issue #39 <https://github.com/libattachsql/libattachsql/issues/39>`_)
 
 Version 0.1
 -----------